requirements.txt
flask==3.1.0
chromadb==0.6.3
PyPDF2==3.0.1
google-generativeai==0.8.5
python-dotenv==1.1.0
waitress==3.0.2
